Command exit codes weren't showing up
This looks like a change between wheezy and jessie auditing that I
didn't notice.   The exit value for the command is now only in a0
of the exit system call record, and that wasn't getting recorded.
Put the exit status after the command, on STOP.

Also document in the man page that the audit subsystem doesn't generate
audit records for processes terminated by uncaught signals, so STOP
account records won't be generated.