Check for ANOM_ABEND also, since no exit call
Ticket: CM-12368
Reviewed By: olson
Testing Done: built, ran process that segv's

ANOM_ABEND is generated for at least some processes that segv, etc.
It's not generated always, but getting some of them is worth it.
Show exitsig= rather than exit= in the accounting record.

The ANOM_ABEND event does not get flushed to our plugin until a
following event, and when it's flushed, it has the timestamp of
the next event.  I can't find a way to change that.

Also fixed the rather strange indenting I had in handle_event()
Use the -b option to git show or diff to not see it.
I fixed up a few comments as well.