Add capability to run within a management VRF network
When running in an environment with the management network setup to
run as a VRF, the audisp-tacplus isn't able to connect to the TACACS+
accounting servers unless it is also configured to run in a VRF.

Added a non-default vrf=NAME config file item and processing to enable
running within the VRF management network.  We only try to set our vrf when
we try to connect to the servers to send accounting records.  Keep
trying (if configured) until we connect successfully.

If we have connected in the past, and now fail for all, and have vrf
configured, assume networking may have restarted, and set the vrf again
and retry the connection.

Fix bug where I wasn't resetting our config state when sent SIGHUP,
before re-reading the config file.  Fixed that, also clear our vrf
connection state.

Fixed bug in return status of tac_connect_single, improved syslog mesg

Document the SIGHUP handling to reread the config file.
Document the config file, rather than pointing to the source

Updated changelog