Add missing config variables to config reset, reorder
Noticed that we were missing resetting some variables here, and that
there was a very narrow window where we might access freed memory.

set the number of servers to 0 before freeing the memory, saving
the previous value in a temp variable.

That could potentially account for the coredump during SIGHUP processing.